Remove themes/ submodules
The themes/ directory provides Hugo with stylistic themes in order to display content. Currently, the theme inside is a Git submodule, which is proving complicated to manage.

In order to allow this to be built locally, reliably, instead remove the themes/ directory, and its children, and bring the theme in to the codebase directly.

This might not be brilliant Git-fu, but it is much easier to manage.