Remove KeyName from CloudFormation template
We don't need a key to be inserted to the instance on launch because we're not going to need access to the instance. If we mandate the use of a key name here, we have to manually create one since CloudFormation doesn't suport creating one for us directly. Another approach would be to create a custom resource, but - honestly - I can't be bothered.

Removes the KeyName and references.